@@ -13,17 +13,17 @@
   "private": true,
   "dependencies": {
     "@angular-redux/store": "^9.0.0",
-    "@angular/animations": "^7.0.4",
+    "@angular/animations": "^8.2.3",
     "@angular/cdk": "^6.4.7",
-    "@angular/common": "^7.0.4",
-    "@angular/compiler": "^7.0.4",
-    "@angular/core": "^7.0.4",
-    "@angular/forms": "^7.0.4",
+    "@angular/common": "^8.2.3",
+    "@angular/compiler": "^8.2.3",
+    "@angular/core": "^8.2.3",
+    "@angular/forms": "^8.2.3",
     "@angular/material": "^6.4.7",
-    "@angular/platform-browser": "^7.0.4",
-    "@angular/platform-browser-dynamic": "^7.0.4",
-    "@angular/platform-server": "^7.0.4",
-    "@angular/router": "^7.0.4",
+    "@angular/platform-browser": "^8.2.3",
+    "@angular/platform-browser-dynamic": "^8.2.3",
+    "@angular/platform-server": "^8.2.3",
+    "@angular/router": "^8.2.3",
     "adm-zip": "^0.4.13",
     "ajv": "^6.5.5",
     "angular-in-memory-web-api": "^0.8.0",
@@ -33,20 +33,20 @@
     "hoek": "^5.0.4",
     "jquery": "^3.3.1",
     "ng2-izitoast": "^2.0.0",
-    "ngx-bootstrap": "^3.1.1",
+    "ngx-bootstrap": "^5.1.2",
     "popper.js": "^1.14.5",
     "redux": "^4.0.1",
-    "rxjs": "^6.3.3",
+    "rxjs": "^6.5.2",
     "timespan": "github:caddis/timespan",
     "url-parse": "^1.4.4",
     "webpack-dev-server": "^3.1.14",
     "zone.js": "~0.9.1"
   },
   "devDependencies": {
-    "@angular-devkit/build-angular": "~0.802.0",
-    "@angular/cli": "^8.2.0",
-    "@angular/compiler-cli": "^7.0.4",
-    "@angular/language-service": "^7.0.4",
+    "@angular-devkit/build-angular": "~0.803.0",
+    "@angular/cli": "^8.3.0",
+    "@angular/compiler-cli": "^8.2.3",
+    "@angular/language-service": "^8.2.3",
     "@types/jasmine": "^3.0.0",
     "@types/jasminewd2": "^2.0.6",
     "@types/node": "^10.12.9",
@@ -65,6 +65,6 @@
     "rxjs-tslint": "^0.1.5",
     "ts-node": "^7.0.1",
     "tslint": "^5.11.0",
-    "typescript": "3.1.6"
+    "typescript": "3.5.3"
   }
 }
